add mode change app icon + set default app icon

+++ features/PostInstallation.js
@@ -1,4 +1,6 @@
+const HandleIcon = require("../icon-utils/HandleIcon");
 const CustomPromise = require("../promises");
+const Texts = require("../texts");
 
 const isWinOS = process.platform === "win32";
 
@@ -9,25 +11,31 @@ const handleRunSimulatorIOS = async (appName) => {
     );
 };
 
-const handleExec = async (appName) => {
+const handleExec = async (appName, repoURL) => {
     // Delete README to avoid conflict and then pull master
-    await CustomPromise.execCommandLinePromise(
-        `cd ${appName} && rm -rf README.md`
-    );
-    await CustomPromise.execCommandLinePromise(
-        `cd ${appName} && git pull origin master --allow-unrelated-histories`
-    );
+    if (repoURL) {
+        await CustomPromise.execCommandLinePromise(
+            `cd ${appName} && rm -rf README.md`
+        );
+        await CustomPromise.execCommandLinePromise(
+            `cd ${appName} && git pull origin master --allow-unrelated-histories`
+        );
+    }
+    // Change app icon to default AMELA icon
+    const appNameWithoutHyphen = `${appName.trim().replace(/-/g, "").replace(/ /g, "")}`;
+    await HandleIcon.generateAndInstallIcons(appNameWithoutHyphen, appName);
+
     if (!isWinOS) {
         // Ask user what to do next
         console.log("Installation completed!");
         const postInstallQuestion = "What do you want to do next?";
         const postInstallAnswerObj = await CustomPromise.getRadioButtonAnswerPromise(
             postInstallQuestion,
-            ["Run on iOS simulator", "Nothing"]
+            [Texts.runOnIOSSimulator, Texts.nothing]
         );
         const postInstallAnswer = postInstallAnswerObj[postInstallQuestion];
         // Running on device iOS
-        if (postInstallAnswer === "Run on iOS simulator") {
+        if (postInstallAnswer === Texts.runOnIOSSimulator) {
             await handleRunSimulatorIOS(appName);
         }
     }
